Changes logic of solvepuzzle.m to a simpler format
Now rather than going through and tediously trying to replicate my own solution techniques, this simply removes
from the pencil array of each box the numbers in the row and column and the 'sub-box' (the three-by-three) that
the cell is in. This doesn't work as far as solving the puzzle goes but it does run so I am committing it. The
next commit will add the further required logic that may result in a working solver.